+++ modules/ads/genrevibes_ads_admob/lib/genrevibes_ads_admob.dart
@@ -4,3 +4,4 @@ library;
 export 'src/admob_ad_provider.dart';
 export 'src/admob_client.dart';
 export 'src/admob_configuration.dart';
+export 'src/admob_test_ads.dart';